Clean the src folders to refactor the components and, to start afresh with login component
