+++ e-keithel-backend/config/routes.rb
@@ -1,5 +1,12 @@
 Rails.application.routes.draw do
-  devise_for :users, controllers: {
+  devise_for :users, 
+  path: '', 
+  path_names: {
+    sign_in: 'login',
+    sign_out: 'logout',
+    registration: 'signup'
+  },
+  controllers: {
     sessions: 'users/sessions',
     registrations: 'users/registrations',
   }
