+++ e-keithel-frontend/src/App.js
@@ -1,16 +1,17 @@
 import { Routes, Route } from 'react-router-dom';
-import Register from './components/Register';
-import Login from './components/Login';
-import Layout from './components/Layout';
-import Home from './components/Home';
-import Product from './components/Product';
-import Cart from './components/Cart';
-import Unauthorized from './components/Unauthorized';
-import Linkpage from './components/Linkpage';
-import Checkout from './components/Checkout';
-import Profile from './components/Profile';
-import History from './components/History';
-import RequireAuth from './components/RequireAuth';
+
+import Register from './app/components/auth/Register';
+import Login from './app/components/auth/Login';
+import Layout from './app/layouts/Layout';
+import Home from './app/components/Home';
+import Product from './app/components/Product';
+import Cart from './app/components/Cart';
+import Unauthorized from './app/components/Unauthorized';
+import Linkpage from './app/components/Linkpage';
+import Checkout from './app/components/Checkout';
+import Profile from './app/components/Profile';
+import History from './app/components/History';
+// import RequireAuth from './components/RequireAuth';
 
 // const ROLES = {
 //   User: 2001,
@@ -24,9 +25,9 @@ function App() {
       <Routes>
         <Route path="/" element={<Layout />}>
           {/* public routes */}
+          <Route index element={<Home />} />
           <Route path="login" element={<Login />} />
           <Route path="register" element={<Register />} />
-          <Route path="home" element={<Home />} />
           <Route path="products" element={<Product />} />
           <Route path="cart" element={<Cart />} />
           <Route path="linkpage" element={<Linkpage />} />
