+++ e-keithel-frontend/src/app/components/RequireAuth.js
@@ -1,0 +1,32 @@
+// import { useLocation, Navigate, Outlet } from 'react-router-dom';
+// import useAuth from '../hooks/useAuth';
+
+// const RequireAuth = ({ allowedRoles }) => {
+//   console.log('allowedRoles', allowedRoles);
+//   const { auth } = useAuth();
+//   const location = useLocation();
+
+//   // return auth?.user ? (
+//   //   <Outlet />
+//   // ) : (
+//   //   <Navigate to="/login" state={{ from: location }} replace />
+//   // );
+
+//   return auth?.roles?.find((role) => allowedRoles?.includes(role)) ? (
+//     <Outlet />
+//   ) : auth?.user ? (
+//     <Navigate to="/unauthorized" state={{ from: location }} replace />
+//   ) : (
+//     <Navigate to="/login" state={{ from: location }} replace />
+//   );
+// };
+
+// export default RequireAuth;
+
+import React from 'react';
+
+const RequireAuth = () => {
+  return <div>RequireAuth</div>;
+};
+
+export default RequireAuth;
