--- e-keithel-frontend/src/components/RequireAuth.js
@@ -1,24 +1,0 @@
-import { useLocation, Navigate, Outlet } from 'react-router-dom';
-import useAuth from '../hooks/useAuth';
-
-const RequireAuth = ({ allowedRoles }) => {
-  console.log('allowedRoles', allowedRoles);
-  const { auth } = useAuth();
-  const location = useLocation();
-
-  // return auth?.user ? (
-  //   <Outlet />
-  // ) : (
-  //   <Navigate to="/login" state={{ from: location }} replace />
-  // );
-
-  return auth?.roles?.find((role) => allowedRoles?.includes(role)) ? (
-    <Outlet />
-  ) : auth?.user ? (
-    <Navigate to="/unauthorized" state={{ from: location }} replace />
-  ) : (
-    <Navigate to="/login" state={{ from: location }} replace />
-  );
-};
-
-export default RequireAuth;
