+++ e-keithel-frontend/src/index.css
@@ -1,4 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Nunito&display=swap');
+/* @import url('https://fonts.googleapis.com/css2?family=Nunito&display=swap');
 
 * {
   margin: 0;
@@ -120,4 +120,4 @@ button {
   display: flex;
   justify-content: flex-start;
   align-items: flex-end;
-}
+} */
